# [IBD] multi-byte block obfuscation