--- a/src/node/mempool_persist.cpp
+++ b/src/node/mempool_persist.cpp
@@ -60,15 +60,15 @@ bool LoadMempool(CTxMemPool& pool, const fs::path& load_path, Chainstate& active
     try {
         uint64_t version;
         file >> version;
-        std::vector<std::byte> xor_key;
+        std::vector<std::byte> obfuscation;
         if (version == MEMPOOL_DUMP_VERSION_NO_XOR_KEY) {
             // Leave XOR-key empty
         } else if (version == MEMPOOL_DUMP_VERSION) {
-            file >> xor_key;
+            file >> obfuscation;
         } else {
             return false;
         }
-        file.SetXor(xor_key);
+        file.SetObfuscation(obfuscation);
         uint64_t total_txns_to_load;
         file >> total_txns_to_load;
         uint64_t txns_tried = 0;
@@ -180,12 +180,12 @@ bool DumpMempool(const CTxMemPool& pool, const fs::path& dump_path, FopenFn mock
         const uint64_t version{pool.m_opts.persist_v1_dat ? MEMPOOL_DUMP_VERSION_NO_XOR_KEY : MEMPOOL_DUMP_VERSION};
         file << version;
 
-        std::vector<std::byte> xor_key(Obfuscation::KEY_SIZE);
+        std::vector<std::byte> obfuscation(Obfuscation::KEY_SIZE);
         if (!pool.m_opts.persist_v1_dat) {
-            FastRandomContext{}.fillrand(xor_key);
-            file << xor_key;
+            FastRandomContext{}.fillrand(obfuscation);
+            file << obfuscation;
         }
-        file.SetXor(xor_key);
+        file.SetObfuscation(obfuscation);
 
         uint64_t mempool_transactions_to_write(vinfo.size());
         file << mempool_transactions_to_write;