@l0rinc 2024-10-31
> Taking a step back, I wonder if this is worth it. IIRC it gives a +1% speedup 

That's not the main point, rather that we're storing the key in a value that can be short-circuited easily so that when key is 0 (i.e. xor is NOOP), we can skip it. Previously this would have only been possible by checking each byte of the key.
It's also a lot cleaner to store it in a primitive instead, which supports xor natively.
Xor comes up in every profiling I do, we shouldn't have a regression because of https://github.com/bitcoin/bitcoin/pull/28207 - this PR solves that.