@l0rinc 2024-10-31
> change a "fix" is not entirely accurate

when xor is disabled we're not xor-ing now at all. Previously we did xor, so this change brings back the previous behavior when xor is disabled.

> The pull request title starts with "optimization", so I got the impression that a speedup is the main point.

Yes, please see the updated description with the benchmarks: https://github.com/bitcoin/bitcoin/pull/31144#issue-2610689777

> may not be visible at all if IO is completely taken out of the critical path

I'd argue the current implementation is slightly simpler (i.e. xor is stored and performed natively and can be disabled) and faster (2x for a representative dataset).