@l0rinc 2024-11-02
Thanks for the hints @maflcko, I was under the impression that big-endian tests were run automatically.

### Fix

It seem that that `std::rotr` doesn't take endianness into account, so the fix basically looks like:
```C++
size_t key_rotation = 8 * key_offset;
if constexpr (std::endian::native == std::endian::big) key_rotation *= -1;
return std::rotr(key, key_rotation);
```
I've emulated the s390x behavior locally like this:

<details>
<summary>Details</summary>

```bash
brew install podman pigz
softwareupdate --install-rosetta
podman machine init
podman machine start
docker run --platform linux/s390x -it ubuntu:22.04 /bin/bash
    apt update && apt install -y build-essential libtool pkg-config bsdmainutils python3 libevent-dev libboost-system-dev libboost-filesystem-dev libboost-chrono-dev libboost-test-dev libboost-thread-dev libsqlite3-dev libminiupnpc-dev libzmq3-dev libssl-dev libgmp-dev libdb-dev libdb++-dev ccache cmake && \
    cd /mnt && git clone https://github.com/bitcoin/bitcoin.git && cd bitcoin && git remote add l0rinc https://github.com/l0rinc/bitcoin.git && git fetch --all && git checkout l0rinc/optimize-xor && \
    cmake -B build && cmake --build build --target test_bitcoin -j$(nproc) && \
    ./build/src/test/test_bitcoin --run_test=streams_tests
```

</details>


### Changes

* The change also includes an updated benchmarking suite with 700k blocks inspected for all usages of `utils::Xor` to make the benchmark representative:
<img src="https://github.com/user-attachments/assets/83f16fc6-1596-40af-a813-e4e67982a051">

* Changed `AutoFileXor` benchmark to measure the regression of turning off obfuscation.

* I've also updated all key derivations to vector-to-uint64 instead of generating the 64 bit key directly (it's more consistent, more representative and helps with endianness).

* Added a xor roundtrip test which applies the xor in random chunks, asserts that it differs from the original and reapplies it in different random chunks and asserts that it's equivalent to the original.

See the complete diff: https://github.com/bitcoin/bitcoin/compare/a5cad729c76cafa047a2b1897595669ae9b2b0d5..af778c5bdea4175f84e82b41f3b51c7f453d8c7e