@l0rinc 2024-11-05
Updated the benchmark to 860k blocks (September 2024):
<img src="https://github.com/user-attachments/assets/1d70f5d9-68cb-4e5f-85e9-ffb9fd3b3f1c">

This one contains a lot of very big arrays (96'233 separate sizes, biggest was 3'992'470 bytes long) - a big departure from the previous 400k and 700k blocks (having 1500 sizes, biggest was 9319 bytes long). 

The performance characteristics are also quite different, now that we have more and bigger byte arrays:

> C++ compiler .......................... AppleClang 16.0.0.16000026

> Before:

|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                1.29 |      774,577,944.12 |    0.2% |    115.99 | `XorHistogram`

> After:

|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                0.04 |   26,411,646,837.32 |    0.2% |      8.97 | `XorHistogram`

i.e. ~35x faster with Clang at processing the data with representative histograms.

> C++ compiler .......................... GNU 13.2.0

> Before:

|             ns/byte |              byte/s |    err% |        ins/byte |        cyc/byte |    IPC |       bra/byte |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|                0.97 |    1,032,916,679.87 |    0.0% |            9.01 |            3.29 |  2.738 |           1.00 |    0.0% |     86.58 | `XorHistogram`

> After:

|             ns/byte |              byte/s |    err% |        ins/byte |        cyc/byte |    IPC |       bra/byte |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|                0.10 |   10,369,097,976.62 |    0.0% |            0.32 |            0.33 |  0.985 |           0.06 |    0.6% |      8.63 | `XorHistogram`

i.e. ~10x faster with GCC at processing the data with representative histograms.