@andrewtoth 2025-01-11
I benchmarked this PR rebased onto 35bf426e02210c1bbb04926f4ca2e0285fbfcd11 up to block 878k two times each, and I saw a 9% speedup. This was using `dbcache=30000`.

```
hyperfine --export-markdown ~/bench.md --show-output --parameter-list commit 1298bae74a1f690fd6cc0e029e490537cbeb301b,35bf426e02210c1bbb04926f4ca2e0285fbfcd11 --prepare 'git checkout {commit} && \
cmake --build build -j $(nproc); \
rm -rf /home/user/.bitcoin; \
sync; sudo /sbin/sysctl vm.drop_caches=3;' -M 2 'echo '{commit}' && /usr/bin/time ./build/src/bitcoind -printtoconsole=0 -dbcache=30000 -connect=local_node -stopatheight=878000'
```

This branch:
```
  Time (mean ± σ):     17147.622 s ± 37.933 s    [User: 35867.007 s, System: 871.738 s]
  Range (min … max):   17120.799 s … 17174.444 s    2 runs
```

Master branch commit 35bf426e02210c1bbb04926f4ca2e0285fbfcd11:
```
  Time (mean ± σ):     18704.258 s ±  4.425 s    [User: 37215.600 s, System: 891.293 s]
  Range (min … max):   18701.130 s … 18707.387 s    2 runs
```
```
Summary
  'echo 1298bae74a1f690fd6cc0e029e490537cbeb301b && /usr/bin/time ./build/src/bitcoind -printtoconsole=0 -dbcache=30000 -connect=192.168.2.171 -stopatheight=878000' ran
    1.09 ± 0.00 times faster than 'echo 35bf426e02210c1bbb04926f4ca2e0285fbfcd11 && /usr/bin/time ./build/src/bitcoind -printtoconsole=0 -dbcache=30000 -connect=192.168.2.171 -stopatheight=878000'
```